requirements.txt
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
transformers>=4.30.0
imapclient>=2.3.1
python-dotenv>=1.0.0
tqdm>=4.65.0
pandas>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0